+++ homework_arch/homework_04_string_regex/js/tasks/task-6.js
