+++ homework_arch/homework_04_string_regex/js/utils/truncate-by-dots/truncate-by-dots.js
