+++ homework_arch/homework_04_string_regex/js/utils/validate-cyrillic-full-name/validate-cyrillic-full-name.js
