+++ homework_arch/homework_04_string_regex/js/utils/validate-document-id/validate-document-id.js
