+++ homework_arch/homework_04_string_regex/scss/footer.scss
@@ -1,0 +1,8 @@
+.footer {
+  min-height: 25px;
+  padding-top: 2px;
+  font-size: 18px;
+  font-weight: 400;
+  text-align: right;
+  background-color: rgb(235, 235, 235);
+}
