+++ homework_arch/homework_04_string_regex/scss/global.scss
@@ -1,0 +1,27 @@
+html,
+body {
+  box-sizing: border-box;
+  padding: 0;
+  margin: 0;
+  font-family: "Roboto", "Arial", sans-serif;
+  font-size: 16px;
+  font-style: normal;
+
+  font-weight: 400;
+  line-height: 1.2;
+  color: rgb(20, 20, 20);
+
+  background-color: rgb(245, 245, 245);
+}
+
+body {
+  min-width: 760px;
+  max-width: 1200px;
+  margin: 0 auto;
+}
+
+*,
+::after,
+::before {
+  box-sizing: inherit;
+}
