+++ homework_arch/homework_04_string_regex/scss/header.scss
@@ -1,0 +1,7 @@
+.header {
+  min-height: 40px;
+  font-size: 30px;
+  font-weight: 700;
+  text-align: center;
+  background-color: rgb(235, 235, 235);
+}
