+++ homework_arch/homework_04_string_regex/scss/main.scss
@@ -1,0 +1,14 @@
+//noinspection CssUnknownTarget
+@import "normalize";
+//  main
+@import "global";
+//  base
+@import "header";
+@import "tasks";
+@import "task";
+@import "start-button";
+@import "err";
+@import "area-name";
+@import "text-area";
+@import "input-area";
+@import "footer";
