+++ src/scss/main.scss
@@ -4,11 +4,4 @@
 @import "global";
 //  base
 @import "header";
-@import "tasks";
-@import "task";
-@import "start-button";
-@import "err";
-@import "area-name";
-@import "text-area";
-@import "input-area";
 @import "footer";
